+++ tests/issues/test_issue_40.py
@@ -4,11 +4,19 @@ https://github.com/MAK-Relic-Tool/Issue-Tracker/issues/40
 """
 
 import io
+from argparse import ArgumentError
 from contextlib import redirect_stderr
 from pathlib import Path, PurePath
 from typing import Iterable
 
 import pytest
+
+
+def _ArgumentError(name: str, message: str):
+    _ = ArgumentError(None, message)
+    _.argument_name = name
+    return _
+
 
 _F = Path(__file__)
 _DIR = _F.parent
@@ -20,40 +28,45 @@ _PH_SGA = "placeholder.sga"
 _ARGS = [
     (
         ["sga", "pack", "v2", str(_DIR_THAT_DOESNT_EXIST), _PH_SGA],
-        f"error: argument manifest: The given path '{str(_DIR_THAT_DOESNT_EXIST)}' does not exist!",
+        _ArgumentError(
+            "manifest",
+            f"The given path '{str(_DIR_THAT_DOESNT_EXIST)}' does not exist!",
+        ),
     ),
     (
         ["sga", "pack", "v2", str(_DIR), "dummy"],
-        f"error: argument manifest: The given path '{str(_DIR)}' is not a file!",
+        _ArgumentError("manifest", f"The given path '{str(_DIR)}' is not a file!"),
     ),
     (
         ["sga", "pack", "v2", str(_DIR)],
-        f"error: argument manifest: The given path '{str(_DIR)}' is not a file!",
+        _ArgumentError("manifest", f"The given path '{str(_DIR)}' is not a file!"),
     ),
     (
         ["sga", "pack", "v2", str(_ILLEGAL_PATH)],
-        f"error: argument manifest: The given path '{str(_ILLEGAL_PATH)}' does not exist!",
+        _ArgumentError(
+            "manifest", f"The given path '{str(_ILLEGAL_PATH)}' does not exist!"
+        ),
     ),
 ]
 _ARGS2 = [
-    ([av.replace("\\", "/") for av in a], str(PurePath(b.replace("\\", "/"))))
+    (
+        [av.replace("\\", "/") for av in a],
+        b,
+    )  # _ArgumentError(b.argument_name,b.message.replace("\\", "/")))
     for (a, b) in _ARGS
 ]
 
 _ = _ARGS2
 
 
-@pytest.mark.parametrize(["args", "msg"], [*_ARGS, *_ARGS2])
-def test_argparse_error(args: Iterable[str], msg: str):
+@pytest.mark.parametrize(["args", "err"], [*_ARGS, *_ARGS2])
+def test_argparse_error(args: Iterable[str], err: ArgumentError):
     from relic.core.cli import CLI
 
-    with io.StringIO() as f:
-        with redirect_stderr(f):
-            status = CLI.run_with(*args)
-            assert status == 2
-        f.seek(0)
-        err = f.read()
-
-        if msg not in err:  # Dumb, but helps avoid bloat
-            print(err)
-        assert msg in err
+    try:
+        _ = CLI.run_with(*args)
+    except ArgumentError as exc:
+        assert exc.argument_name == err.argument_name
+        assert exc.message == err.message
+    else:
+        pytest.fail("No Error Raised")
